+++ monkey-tests/tests/fuzz_multi_pool.rs
@@ -13,5 +13,5 @@ fn fuzz_multi_pool() {
         }
     }
 
-    FuzzTest::<MultiResourcePoolFuzzer>::run_fuzz(32, 100);
+    FuzzTest::<MultiResourcePoolFuzzer>::run_fuzz(32, 100, false);
 }
